Primeros cambios

+++ Models/ActividadCliente.cs
@@ -6,9 +6,9 @@ namespace Proyecto.Models
     public partial class ActividadCliente
     {
         public int Id { get; set; }
-        public int Cliente { get; set; }
+        public long Cliente { get; set; }
         public string Descripcion { get; set; } = null!;
-        public int Asesor { get; set; }
+        public long Asesor { get; set; }
 
         public virtual Usuario AsesorNavigation { get; set; } = null!;
         public virtual Cliente ClienteNavigation { get; set; } = null!;
