+++ Models/ActividadContacto.cs
@@ -6,9 +6,9 @@ namespace Proyecto.Models
     public partial class ActividadContacto
     {
         public int Id { get; set; }
-        public int Contacto { get; set; }
+        public long Contacto { get; set; }
         public string Descripcion { get; set; } = null!;
-        public int Asesor { get; set; }
+        public long Asesor { get; set; }
 
         public virtual Usuario AsesorNavigation { get; set; } = null!;
         public virtual Contacto ContactoNavigation { get; set; } = null!;
