+++ Models/ActividadCotizacion.cs
@@ -8,7 +8,7 @@ namespace Proyecto.Models
         public int Id { get; set; }
         public int Cotizacion { get; set; }
         public string Descripcion { get; set; } = null!;
-        public int Asesor { get; set; }
+        public long Asesor { get; set; }
 
         public virtual Usuario AsesorNavigation { get; set; } = null!;
         public virtual Cotizacion CotizacionNavigation { get; set; } = null!;
