+++ Models/ActividadEjecucion.cs
@@ -8,7 +8,7 @@ namespace Proyecto.Models
         public int Id { get; set; }
         public int Ejecucion { get; set; }
         public string Descripcion { get; set; } = null!;
-        public int Asesor { get; set; }
+        public long Asesor { get; set; }
 
         public virtual Usuario AsesorNavigation { get; set; } = null!;
         public virtual Ejecucion EjecucionNavigation { get; set; } = null!;
