+++ Models/CotizacionCliente.cs
@@ -7,8 +7,8 @@ namespace Proyecto.Models
     {
         public int Id { get; set; }
         public int Cotizacion { get; set; }
-        public int Asesor { get; set; }
-        public int Cliente { get; set; }
+        public long Asesor { get; set; }
+        public long Cliente { get; set; }
 
         public virtual Usuario AsesorNavigation { get; set; } = null!;
         public virtual Cliente ClienteNavigation { get; set; } = null!;
