+++ Models/TareaCaso.cs
@@ -10,7 +10,7 @@ namespace Proyecto.Models
         public string Descripcion { get; set; } = null!;
         public DateTime? FechaFinalizacion { get; set; }
         public int EstadoTarea { get; set; }
-        public int Asesor { get; set; }
+        public long Asesor { get; set; }
 
         public virtual Usuario AsesorNavigation { get; set; } = null!;
         public virtual Caso CasoNavigation { get; set; } = null!;
